+++ tests/testthat/test-syllable_sum.R
@@ -19,8 +19,8 @@ test_that("syllable_count, gives the desired output",{
             "lie", "like", "robots"), class = "factor"), syllables = c(2, 
             1, 2, 1), in.dictionary = structure(c(1L, 1L, 2L, 1L), .Label = c("-", 
             "NF"), class = "factor")), .Names = c("words", "syllables", 
-        "in.dictionary"), row.names = c(NA, -4L), class = "data.frame")), .Names = c("ALGORITHM REPORT", 
-        "SYLLABLE DATAFRAME"))
+        "in.dictionary"), row.names = c(NA, -4L), class = "data.frame")), 
+        .Names = c("ALGORITHM REPORT", "SYLLABLE DATAFRAME"))
    
     expect_true(identical(x1, x1_c))
     expect_true(identical(x2, x2_c))
@@ -35,19 +35,16 @@ test_that("syllable_sum, gives the desired output",{
         "syllable_freq", "numeric"), wc = c(6L, 5L, 4L, 4L, 5L, 5L, 4L, 
         3L, 5L, 6L, 6L), type = "Syllable")
 
-
     expect_true(all(x3 == x3_c))
     
 })
 
 test_that("polysyllable_sum, gives the desired output",{
     
-    x4 <- polysyllable_sum(DATA$state)
-    
-    x4_c <- structure(c(1L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 1L), class = c("polysyllable_sum", 
-        "syllable_freq", "integer"), wc = c(6L, 5L, 4L, 4L, 5L, 5L, 4L, 
-        3L, 5L, 6L, 6L), type = "Pollysyllable")
-   
+    x4_c <- structure(c(1L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 1L), 
+        class = c("polysyllable_sum", "syllable_freq", "integer"), 
+        wc = c(6L, 5L, 4L, 4L, 5L, 5L, 4L, 3L, 5L, 6L, 6L), 
+        type = "Pollysyllable")
 
     expect_true(all(x4 ==  x4_c))
     
@@ -65,4 +62,15 @@ test_that("combo_syllable_sum, gives the desired output",{
 
     expect_true(identical(x5, x5_c))
     
+})
+
+test_that("cumulative methods for syllable_freq, gives the desired output",{
+ 
+    x3_cum <- cumulative(syllable_sum(DATA$state))
+    x4_cum <- cumulative(polysyllable_sum(DATA$state)  )
+    
+    expect_true(is.data.frame(x3_cum))
+    expect_true(all(colnames(x3_cum) == c("cumave", "Time")))
+    expect_true(is.data.frame(x4_cum))
+    expect_true(all(colnames(x4_cum) == c("cumave", "Time")))
 })